More narrow scope for sequencer timeline drag events
A side benefit of the layout changes introduced in commit
908c35b8375ad53b450a46f562b42159e6d2209b is that they allow moving
the event handlers that deal with dragging pattern boxes to a more
specific DOM element. This allows simplifying a calculation, and
allows for an simpler mental model since it no longer has to
account for elements that you can't drag into.