Fixing dashed line appearance in pattern boxes
The previous implementation caused the dashed vertical line that
is shown in a looped pattern to look different depending on the
browser/platform. In some cases it looked obviously bad (too
thick/thin, or misaligned). The issue seems to have been the 0.5px
border width, as different browsers/platforms seem to display it
differently. However, a 1px border seems to be displayed
consistently.