Preventing sequencer popup menu content wrapping
Previously, when a sequencer popup menu was triggered too close to
the window's right edge the contents of the popup menu could be
wrapped to multiple lines. This caused the popup menu to be in the
wrong position vertically, and also caused the inner contents to
look bad.

The problem was that the popup menu's inner contents can have a
fractional width, but the width being used for determining the
menu position was a rounded integer, due to using `.clientWidth`.
When the menu was clamped to the right edge of the screen, and the
width was rounded such that it rounded down, the calculation used
for positioning could cause the popup menu to be shrunk
horizontally ever so slightly, causing the contents to wrap.

The solution is to use `.getBoundingClientRect()` to determine
the width, since it returns the true fractional dimensions.