Consistent mouse/touch behavior for showing popup menu
Previously, one tap on a sequencer pattern box with a touch event
would cause the box to be highlighted and the popup menu to be
shown. However, showing the popup menu required two clicks with
the mouse - one to highlight the pattern box, and a further one to
show the popup menu.

Now the behavior is consistent for both touch and mouse, in favor
of the previous mouse behavior. (I.e., two clicks/taps to show the
menu). This was the intended behavior all along, but was not the
case due to a bug.