Sounds files list for song dynamically generated
Previously the list of sound files to load when loading the
default song was hard-coded. Now it is dynamically generated from
the list of instruments in the song configuration. This removes
the need to change data in two different places when changing the
default song, and is a step toward making it possible to load new
songs at runtime.