Refactoring how default song data is added to app state
The initial app state no longer contains hard-coded values
specific to the default song. Instead, the relevant property
values are all defined in the `DefaultSong` object and applied to
the app state in the `initialize()` function. This incrementally
moves the code toward the ability to load different songs at
runtime, instead of being hard-coded to load a specific song.

In particular, the measure count and transport tempo field values
are now defined on the `DefaultSong` object, instead of the
initial app state.