Better UI for moving sequencer pattern via dragging
Previously, moving a pattern worked via a simple drag. Whenever
the pattern was dragged to a new location it was immediately
moved. If the user attempted to drag the pattern to a place where
it couldn't be moved (because it would overlap with another
pattern) nothing would happen - the pattern would not be moved,
and no information would be given to the user.

Now, moving a pattern works more like drag-and-drop. When a drag
is started, a transparent "ghost" pattern is shown to indicate
where the pattern will be moved to. The pattern is not actually
moved until it is "dropped". If the pattern is dragged to a place
where it can't be dropped it turns red to alert the user. After
trying it out this seems like a better UI.

Note that this only applies to moving a pattern, not changing its
length or looping it.

A known issue is that if the mouse/touch is dragged outside the
bounds of the timeline grid the drop is not cancelled properly.
This should be fixed in a future commit.