Refactoring logic for checking pattern operation bounds
I.e. if there's enough room at a given spot in a particular track
to move/resize/loop/paste a pattern.

The previous code used a method of optimistically updating the
pattern as if there were room, then checking to see if there's
room, and reverting the changes if not.

The new code checks to see if there's room, and if so, only then
makes changes to the pattern. This is simpler and more succinct.

The core of this is the `isSpaceForPatternInTrack()` function,
which is re-purposed from the previous `isPatternMoveable()`.