Fixing 1 pixel cutout in pattern sidebar
If a pattern name was long enough to be truncated it sometimes
overlapped the pattern sidebar (i.e. the place with the "loop" and
"resize" buttons) by 1 pixel. This resulted in a small cutout on
the left edge of the sidebar.

This was introduced in b980dadab8b2386cbe5c83a4e0004513c0ff70cc.
When the width of each pattern box was increased by 1 pixel, the
width of the pattern name element should have also been changed
by 1 pixel to compensate.