Adding missing argument to function call
I don't think the missing argument caused any problems in
practice, although by accident. If the envelope were in the
release phase the function would have returned the wrong result
due to the missing argument, but in the one place this function
is called the relevant envelope will never be in the release
phase. However, this argument should still be added to avoid
confusion in the future.