Fixing filter envelope release for interactive notes
Specifically, for notes ended before the attack portion of the
envelope is complete. Previously, for certain filter config, such
as high resonance, long attack and release phases, and large
difference between cutoff frequency and attack target frequency,
a quickly ended note could result in a muffled popping sound, with
no release phase of the envelope occuring.

This should be essentially the same fix as the fix in commit
371b437d820bf1754d895128793820dbd73824ea. That commit fixed the
issue for the volume gain envelope, and this commit fixes it for
the filter envelope.

However, although that commit only mentioned fixing things in
Firefox, the issue this commit fixes seemed to exist in Safari,
Firefox and Chrome.