Space under popup menu body allows clicks through
I.e., clicking in the vertical whitespace between the bottom of
the main bubble body and the bottom tip of the arrow will now
allow clicks to pass through. For example, if a sequencer pattern
is selected and the popup menu is shown, this will allow selecting
the pattern above it by clicking in this area. This matches what
you would expect to happen based on the visual appearance.

However, this doesn't work as you would expect when clicking in
the area near the arrow itself. Whitespace in the bounding
rectangle of the arrow does _not_ allow clicks through. Ideally it
would, but I don't know yet how to implement that. Despite that,
this commit still seems like an improvement.