Fixing opacity of dragged sequencer patterns
For some reason, in the minified production build the opacity was
changed from "50%" to "1%" in the main CSS file. My hunch is that
the CSS minifier doesn't support percentage values for opacity,
and clamped the value to the range 0.0 to 1.0. This only occurred
in production builds, not development builds.

After looking into it further, it seems like there is relatively
low browser support for percentage opacity values anyway, so that
is also a good reason to change this.