Removing unnecessary update for React state field
Since the song's measure count can only be changed from the
sequencer, and `selectedPatternID` should always be `undefined`
when the sequencer is displayed, there shouldn't be any reason to
update this field when changing the song's measure count.

This looks like a leftover from the old UI in which the sequencer
and instrument/pattern editors were always shown together in a
unified view.