Fixing cancelled loop drag when using touch events
Previously, looping a pattern on mobile using the "loop" sidebar
button would sometimes stop working mid-drag. Specifically, when
looping caused a pattern segment to be added or removed. To keep
looping, the user had to stop the touch event, and start a new
one.

The problem was that when a segment was added or removed, the
segment containing the sidebar was given a new React key, and
therefore a new DOM element. Since the DOM element that originated
the touch event was gone, the touch event was cancelled.

The solution is to make sure the final segment (i.e. the one with
the sidebar) always has the same key. A simple way to do this is
to give segments keys based on reverse index - this way, the
segment with the sidebar will alway have a key of 0. (Segments
don't have a natural ID that could otherwise be used).