Better behavior when pattern dragged out of timeline
Previously, when a drag-move was in progress and the mouse/touch
moved outside the sequencer timeline bounds, the ghost pattern
would still be displayed where it last was before the mouse/touch
exited. If the mouse re-entered the timeline without the main
button pressed, the pattern would be dropped in the exit location
as well (instead of the drag being cancelled).

Now, the ghost pattern is only shown when the mouse/touch is
within the timeline, and hidden otherwise. In addition, if the
mouse re-enters the timeline after exiting and the main mouse
button isn't pressed, the drag-move will be cancelled. This seems
like more logical behavior.

A known pre-existing issue that hasn't been fixed is that if the
main mouse button is released while the mouse is outside the
timeline, the original pattern will not be restored to full
opacity until the mouse re-enters the timeline.