Fixing bounds checking logic for placing patterns
The logic to determine if there is room in a track to place a
pattern at a particular spot had two off-by-one errors - one
in checking if there is overlap with the beginning of an existing
pattern, and one in checking if there's overlap with the end of
an existing pattern.

This did not actually cause any errors due to patterns being
clamped to measure boundaries by the UI, but this should still be
fixed.