Indication given if not enough room to paste pattern
Previously, the "Paste" button in the sequencer popup menu was
shown in an active state any time there was a pattern copied to
the app's "clipboard". However, if there wasn't enough room to
paste the pattern at the given location, nothing would happen when
you clicked the button. This seemed like confusing behavior.

Now, in this scenario the button will be disabled but with red
styling to indicate that the copied pattern can't be pasted. (As
distinct from the grayed-out disabled state when there is no
pattern copied).