Better field/method names
"Selected" as terminology for these fields made more sense in the
old UI where the sequencer and pattern/instrument editor were
always shown together in a unified view. However, in the current
UI it seems like a confusing word to use, in particular because
it clashes with the notion of clicking on a pattern in the
sequencer to "select" it.

"Being edited" seems like more direct terminology to use. I.e.,
it seems like `patternBeingEditedID` more clearly indicates which
pattern is active in the pattern editor than `selectedPatternID`.