Popup menu shown on tap end, not tap start
Previously, when tapping on a sequencer pattern a popup menu was
shown immediately on mouse down/touch start. Now it is only shown
after the following mouse up/touch end event, and only if the
pattern has not been dragged.

Also, previously the first tap only highlighted the pattern, and
showing the menu required a second tap. Now, the first tap will
show the menu, and the second one will hide it.

A benefit of the show-on-tap-end behavior is that it avoids the
popup menu quickly flashing on and then off when dragging a
pattern.

As for having the first tap show the menu, it feels right because
performing a full tap without dragging seems more intentional than
just starting a tap.

Note that this only affects tapping on patterns, and not tapping
on the underlaying timeline grid itself.