Additional narrower viewports use the "single line" header
The "single line" header is the layout where the logo, transport,
and "download to *.wav" button are all horizontally next to each
other, instead of the transport wrapping to a new line.

This layout is nicer and more compact, and there were some
viewport sizes that had enough room for it but weren't using it.
In particular, this looks better on an iPad in portrait
orientation, especially because it prevents the tempo and volume
sliders from being huge.

To facilitate this, the logo is also made smaller in some narrower
viewport sizes.