Loading screen correctly centered vertically on iOS
In portrait orientation... but not landscape.

The intent is for the loading screen to be exactly the size of the
visible browser content area. However, on iOS, the `vh` unit also
includes the browser chrome (both at the top and bottom of the
screen), while page content starts _below_ the chrome at the top
of the screen. This causes the text to be off center, and page
content to be scrollable.

Using absolute positioning and setting the top and bottom of the
content to be `0` fixes things in portrait orientation. However,
the content is still off center and scrollable in landscape. I'm
currently not sure how to fix this with pure CSS, although maybe
there is a way.