Adjusting logo size on loading screen and app header
The logo in the header is now the same size in all viewport sizes.
It has become smaller in the widest viewport size, and is the same
size as before elsewhere. This makes the logo less "in your face"
while using the app.

Similarly, the logo is now the same size on the loading screen in
all viewport sizes. This is done by making it larger in narrower
viewports. When the app is loading it is the only thing you see,
so it makes more sense for it to be larger.