Avoiding unnecessary pattern preview image re-generations
When a pattern is changed in the pattern editor there is no need
to update its preview image, because it will not be visible.
(Preview images are only visible in the sequencer). It only needs
to be updated when the pattern editor is closed and the sequencer
becomes visible again.