Moving amplitude field from `InstrumentNote` -> `Note`
The amplitude of each note seems like data that is a fundamental
part of the note, and it seems like a historical artifact that it
was part of `InstrumentNote` instead of `Note`.

This change doesn't make a practical difference in the current app
since the amplitude is always hard-coded to `1.0`, but it seems
like a step toward allowing notes to have custom amplitudes in
the future.