Using non-linear sliders for envelope rate fields
This commit introduces a new type of non-linear slider, similar to
the exponential slider introduced in commit
7a0c9e4cc0fd2eb72ddd47b67bd1b6c9b792f0d4. The new slider is
"multi-linear" (for lack of a better term). That is, the value
controlled by the slider always changes linearly, but the slope
differs depending on what section the slider thumb is in. For
example, the range of 0-50% could map to 0-10, while the range
50-100% could map to 10-50.

The benefit of this setup is that it allows the slider to have a
large range, while also enabling fine-tuning for certain sub
ranges. The exponential slider was introduced for the same
purpose, but this new multi-linear slider seems both easier to
use, and easier to configure behind the scenes. For example, it
avoids the problem of allow _too much_ fine-tuning at the low end,
which can be a problem when using an exponential function to map
values.

In this commit, all "attack", "decay", and "sustain" envelope
fields are converted to this new type of slider. This allows more
fine-grained control of time values less than 1 second.