Loading screen displayed more quickly
Previously, a blank white screen was shown until all JavaScript
assets were downloaded and the React app was booted up. Only
_then_ would the loading screen be displayed. Depending on the
Internet connection, this could take a noticeable amount of time.

Now the loading screen will be displayed as soon as the (small)
main HTML file is downloaded. It will then be replaced with an
identical loading screen by the React app once it boots up.
Although this requires duplicating the loading screen in two
different places (i.e. the main HTML file and inside the React
app), the intention is to eventually make the loading screen not
be part of the React app at all. In the meantime, this provides a
better experience for people on a slower connection.