Converting pitch wobble field to multi-linear slider
Instead of an exponential slider.

This allows the slider to devote most of its space (the first 75%)
to wobble amounts between 0 and 100 cents. Previously, the first
75% was devoted to amount between 0 and 675 cents. This allows for
better fine-tuning in the lower end of the range, where the
distinction between different values makes more of a difference.

Since the `<ExponentialSlider>` component is no longer being used
anywhere it is being removed.