Space near popup menu arrow allows clicks through
This completes the work started in commit
1dde3678d2e57d115df965d584693fa6a38013be. In that commit message I
mentioned that I didn't know how to allow clicks to pass through
the negative space in the DOM box containing the popup menu arrow,
but I have since learned about the `clip-path` CSS property.
Clipping the containing DOM box to the same triangular shape as
the arrow causes click events to only be registered when the mouse
is over the arrow itself, but not outside of it.