Preventing disconnected popup menu arrow
Previously, when the target of a popup menu was near the right
edge of the browser window, the main body of the menu was clamped
to the right edge of the browser window, but the arrow underneath
the menu wasn't. This could cause the arrow to become disconnected
from the rest of the menu, which looked weird. (In theory this
could have happened on the left side of the window as well, but in
practice wasn't possible because the sequencer timeline doesn't
extend all the way to the left window edge.

Now the arrow will be clamped to the right side of the popup menu
body, which will prevent the arrow from becoming disconnected.
However, this means that in some (hopefully uncommon) cases the
arrow will not actually point to the actual target point, when the
target is very close to the right edge of the window. Although not
ideal, the connected arrow has the virtue of not looking obviously
broken, and hopefully in context it will make sense what target
the menu corresponds to. The new behavior also seems to match how
similar popup menus work in macOS/iOS when the target is near the
edge of a container.