Simpler method of calculating note frequency
Using a note's MIDI note number to calculate its frequency is
simpler and more succinct. For most notes, the new formula gives
the exact same frequency as the old method. For 19 out of the 96
currently possible notes, there is a very small difference that
should not have any practical effect (as in, the old and new
frequencies don't differ until ~14 decimal places). As a
representative example, for the note G2/43, the old and new
frequencies are:

Old method: 97.99885899543732
New method: 97.99885899543733