Adjusting button border radii
This ties the border radii to the root font size, making the
radius sizes less arbitrary. A motivator of doing this is to allow
reducing the padding of the sequencer popup menu without the
border radius of the menu and the border radius of the menu's
buttons being mismatched in a jarring way.