Preventing spurious drags when using context menu
Previously, right-clicking on a pattern and then selecting an item
from the browser context menu that is displayed would cause the
pattern to be in a bad drag state - an active drag would be in
progress, even though no mouse buttons were being pressed. It
would also usually result in a bunch of errors in the console.

Now, a drag will only begin when the left mouse button is clicked.
The right mouse button still allows the browser context menu to be
displayed, but won't start a drag.

As an alternative I considered disabling the context menu
altogether so that the right mouse button could be used for
dragging. However, I would prefer to keep default browser behavior
where possible, and dragging with the right mouse button seems
like edge case behavior.