Swapping order of "resize" and "loop" buttons
Previously, in each pattern box sidebar the "loop" button was on
top, and the "resize" button was on the bottom. Now the order is
reversed. Although the old ordering matched GarageBand and Logic,
the new ordering seems more logical. (I.e., first you set the size
of the pattern, and then you loop it).