Correct behavior when pattern drag ended outside timeline
Previously, if a pattern drag was in process, and the mouse moved
out of the main timeline element, and then the mouse button was
released while still outside the timeline, the original source
pattern would continue to have a transparent effect even though
the drag was over.

Now, the drag will be correctly ended (and the opaque pattern
appearance restored) regardless of where the mouse button is
when it is released, even if it's outside the browser window.