Hiding popup menu when window is resized
Resizing the browser window can cause the popup menu to become
scrunched against the side of it, causing the inner contents of
the menu to wrap and look weird. The simplest fix seems to be to
hide the menu when the window is resized, similar to how it is
hidden when the timeline is scrolled.