Expanding contents of default song
This expands the length of the song from 8 measures to 16
measures, which allows the song to have two "verse" and "chorus"
sections. Two new tracks are also added, with parts in the 2nd
"verse" section.

These changes allow highlighting some of the recent improvements
to the app, in particular the ability to play longer songs.