Empty measure highlights work like pattern highlights
That is, when clicking on empty space in the sequencer timeline,
the measure highlight is shown immediately upon the "mousedown" or
"touchstart" event, but the popup menu is not shown until the
"mouseup" or "touchend" event. This matches the highlight and
popup menu behavior when clicking on a pattern box. Previously,
the popup menu was shown immediately after the "mousedown" or
"touchstart" event.

Also, the measure highlight and popup menu are now hidden when a
different part of the page is clicked, similar to how pattern
boxes are no longer highlighted when they lose focus.

Although this implementation doesn't appear to have any issues
when used with mouse events, there are two known issues when used
with touch events that will need to be fixed in the future:

1.) Measure highlight is not hidden when timeline is scrolled
2.) Measure highlight is shown after a pattern is removed