Fixing resize/loop pattern buttons for touch events
2618e646a289f7146d730282c385e1f61984f935 accidentally broke the
"resize" and "loop" pattern buttons when using touch events, such
that touching them did nothing.

That commit added a change to make the primary mouse button the
only mouse button that could be used with the "resize" and "loop"
pattern buttons. However, the check for which mouse button was
being pressed assumed that all events have a `button` property,
but this property is undefined for touch events. This caused touch
events to be ignored.