Preventing text selection while dragging patterns
Previously, in Safari text outside the timeline grid could become
selected while dragging a pattern in these scenarios:

* In iOS, by performing a long press
* In macOS, moving the mouse outside the timeline grid while a
  drag was in progress.

This was annoying and never the intended behavior. This did _not_
happen in Firefox or Chrome.

Now all text selection on the page is disabled while a drag is in
progress, and re-enabled when the drag is complete. Although this
approach seems a little bit heavy handed, I'm not currently aware
of a different way to accomplish this.