Fixing pattern loop drag in Firefox
Previously, in Firefox a "loop drag" could be stopped prematurely
while dragging such that the number of loops was decreased. This
did not happen in Safari or Chrome.

This issue was that an in-progress drag was always ended when a
`mouseenter` event was generated on the timeline grid container
and no mouse buttons were pressed. For some reason, Firefox seems
to generate a `mouseenter` event when a new element is added to a
container, and the `buttons` value of the event is always `0`
(i.e., no mouse buttons pressed) regardless of which buttons are
actually pressed. (I _think_ this happens in response to new
elements; I haven't found any documentation online but this seemed
to be the case from testing). Since new elements can be added
during a drag due to React rendering, I believe that is what
caused the offending `mouseenter` events.

The fix is to remove the logic that ends a drag due to a
`mouseenter` event. It turns out that it is no longer necessary,
due to the changes in d353f9931d20ebfa58205cff98b497fdd8e0f42a.