Sequencer popup menu correctly toggled on/off
This commit fixes the popup menu to be correctly toggled on and
off each time a pattern or empty measure in the sequencer is
clicked or tapped. Previously, the popup menu would always be
shown each time a pattern or empty measure was clicked, and could
only be hidden by clicking somewhere outside of the sequencer
timeline, or performing another action that hides the menu such as
scrolling.

As a side effect, this also removes the erroneous empty measure
highlight that was shown after a pattern was removed via the popup
menu using a touch event.

The one exception to the intended behavior is tapping a pattern
with a touch event. For an unknown reason the popup menu is always
shown, instead of being toggled on or off. (This only affects a
pattern box, not an empty measure). That should be fixed in a
future commit.