Popup menu toggled on/off upon repeated touches
When a selected pattern box is tapped multiple times with a touch
event, the popup menu will now be toggled on or off with each tap.
Previously, the popup menu would be shown after every tap. This
makes the behavior consistent with the toggling behavior
elsewhere.

This problem was referred to as an unresolved issue in commit
e5ff92883afedd2dac56bf4c5d3606709dd49794, and it is now resolved.