Hiding measure highlight on scroll/window resize
When the sequencer timeline grid is scrolled or the window is
resized any visible popup menu is hidden. However, previously a
matching measure highlight was _not_ hidden, even though without
the popup menu the highlight didn't make sense. Now the popup menu
and the measure highlight will both be properly hidden in tandem.