Avoiding unnecessary function calls
These function calls should be redundant, because the "loop" and
"resize" buttons on a pattern box are only visible when the
pattern is already selected. I believe these function calls were
an artifact of an older version of the app in which the "loop" and
"resize" buttons were always visible, regardless of whether the
pattern was selected or not.