Avoiding unnecessary pattern box renders
Previously, many clicks or taps on a sequencer pattern box
resulted in all of the other pattern boxes being rendered as well.
However, the re-renderings of the other pattern boxes were usually
unnecessary because they hadn't changed.

The culprits were the `isPopupMenuActive` and `isPopupMenuPending`
props. When a pattern is selected, the values of these props
previously changed for _all_ pattern boxes, causing them to all be
re-rendered (even though the props only had an actual effect on
the currently selected pattern box).

After this commit, `isPopupMenuPending` is defined in a way such
that the value will only change for the currently selected
pattern, and `isPopupMenuActive` is no longer even passed in as a
prop. The end result is that when a pattern box is clicked or
tapped, it will be re-rendered, but the other patterns won't be
unless they actually changed (such as being un-selected).