Preventing text selection for piano keyboard touch events
On iOS, performing a long press selects the text under the touch.
However, text selection is disabled on the piano keyboard, which
causes the closest text adjacent to the keyboard to be selected.
This is annoying, and easy to accidentally trigger. For example,
by holding a long note, or holding down one of the scroll buttons.

This commit prevents this by disabling _all_ text selection on the
page while a touch event is active on the piano keyboard, and
re-enabling text selection when there are no more active touches.
Although this seems a little bit heavy handed, I'm currently not
aware of a different way to accomplish this.

A similar approach could also be useful for disabling text
selection while using the mouse, but this commit only affects
touch events. (Fixing this for mouse events seems to require some
changes to how the mouse events are currently handled).

This is similar to 57d582b4d9da767bc1894213ca44b04945f3e3c8, which
fixed a similar problem for dragging pattern boxes in the
sequencer.