Notes in pattern editor scrolled into view when focused
Previously, when a note box outside the visible editing area was
focused via the keyboard (e.g. using the left/right arrow keys),
it was not automatically scrolled into view. Now when a new note
box is focused, the editing area will always scroll so that it is
visible.

As part of implementing this change, when a note box is focused it
is now actually the focused element on the page (instead of just
appearing like it is focused), and receives keyboard events. This
removes the need for the hidden textbox that previously handled
keyboard events for note boxes by proxy.