Refactoring function for clarity
It seems a little more direct and straightforward to do a simple
iteration of the input array to build up the list of names, rather
than the map+filter approach.

This also removes the unnecessary concatenation of the array with
`[]` in the `setState()` call. This was presumably being done to
make sure the new state value was a different object from the old
value, but since the new value is an array constructed from
scratch, they should always be guaranteed to be different.