More consistent mouse/touch measure highlight behavior
Previously, the behavior of the measure highlight (and popup menu)
that is shown when clicking/tapping on empty space in the timeline
differed slightly for mouse and touch events.

For mouse events, the highlight was shown on mouse down, and the
popup menu was shown on mouse up. A second mouse down would cause
the highlight and popup menu to be hidden.

For touch events, the highlight and popup menu were shown together
on touch end. The next touch start would hide the popup menu, and
the following touch end would hide the measure highlight.

This commit changes touch events consistent with the behavior of
mouse events, because that is the intended behavior.

However, this also introduces a regression in the process. When a
pattern is selected, it now takes an extra touch tap to show the
measure highlight in an empty measure. (This only affects touch
events, not mouse events, and no extra tap is needed if no pattern
is selected). Despite this regression, this commit still seems
like a step in the right direction because it makes the logic
between mouse and touch events more consistent.

As a side effect of these changes, this also fixes a bug in which
clicking directly on a grid line in the sequencer would cause a
measure highlight to be shown, but no popup menu.