New views always initialized scrolled to the top
Previously, the vertical scroll position of the page would be
preserved when switching between the sequencer, instrument editor,
and pattern editor views. However, this could put the new view
into a seemingly arbitrary and unexpected scroll position when it
was first displayed. In particular, this was most noticeable in
smaller viewports, such as on mobile devices.

Now the page's scroll position will always be reset to the top of
the page any time one of the main views is loaded.