Better error appearance for pattern editor note boxes
When a note box in the pattern editor has an invalid value, the
note box now has a thicker red border (instead of a thin gray
border) to make it more consistent with the appearance of a
focused pattern box. The background red color is also now the
standard app-wide red color, instead of a one-off shade of red.