Removing unneeded CSS classes
The `left` and `right` styles are not needed because these
elements are absolutely positioned. As a side effect, the `left`
CSS class can be removed, because this was the only place it was
being used.

The `full-height` style was not needed because the pattern box is
given a specific size in pixels in the `style` attibute.

I'm not sure why the `overflow-hidden` style was there, but there
shouldn't be any overflow to be hidden.