Fixing scroll button highlights for touch events
The left/right scroll buttons on the piano keyboard are shown in a
highlighted style when they are pressed. Previously, the highlight
styles were not correctly shown when both buttons were pressed at
the same time using touch events.

When both buttons were pressed, only the button that was pressed
earlier would show a highlighted style. If the 2nd button was held
and the 1st button was released, the 1st button would continue
to show a highlighted style until the 2nd button was released.
However, touching a key on the keyboard would fix things.

The general problem was that the keyboard was not being
re-rendered when the 2nd button was pressed. This commit fixes
that by basing the presence of the highlight style off of React
state which is now updated when the buttons are pressed or
released.