Fixing margins for views in iPhone X landscape mode
This fixes the pattern and instrument editor views to include
extra padding on the left and right sides in landscape mode on an
iPhone X-style phone. This prevents page content from being cut
off by the rounded phone corners and the notched cutout. For some
reason these views weren't included in the changes in
1e00cf0135e28f13da32b1e0f5ef1249dae6e878.

The piano keyboard in both views still does not contain extra
padding, and can be hidden by the notched cutout. I'm not sure
yet of the best way to fix that from a design perspective, so that
is being left for a future commit.