Popup menu hidden when playback head is dragged
Previously, in iOS an active sequencer popup menu in the
foreground could obscure the playback head while it was being
dragged. This didn't happen in desktop browsers, because the popup
menu was automatically hidden when the playback head slider gained
focus. For some reason, in iOS using a range slider does not seem
to cause a previously focused element to lose focus.

This commit fixes the iOS behavior to match the desktop behavior,
so that an active popup menu is hidden whenever the playback head
slider is dragged.