Better markup formatting to prevent diff noise
When an attribute is added, removed, or modified, ideally only
that attribute should show up in a diff, and not unrelated code.
This new style of formatting allows that to be the case.

For example, previously if a new attribute was added to the end of
a set of attributes, the old final attribute would also show up in
the diff due to it no longer being on the same line as the tag
close character(s). Now that each attribute is on its own line,
that will no longer happen.

In some cases this also seems to improve readability, by breaking
up long lines into several shorter lines.

A downside of this commit is that it will make it harder to use
the Git "blame" history to determine why previous changes were
made. However, if these formatting changes are to be made, the
sooner the better, because the amount of "blame" history that
would be lost will only grow.