Removing obsolete workaround for touch sliders
The code being removed was originally added as a workaround to
allow range sliders inside of a scrollable element to work on iOS.
In older versions of iOS, attempting to change a slider's value
via touch would instead scroll the parent container, making the
slider essentially unusable. Therefore, as a workaround to allow
the playback head slider to be used in iOS, scrolling for the
timeline grid was temporarily disabled while the slider was being
dragged.

As of (I believe) iOS 13, this is no longer the case. Therefore,
this workaround should no longer be necessary.