Preventing invalid octaves in pattern note boxes
Note boxes in the pattern editor prevent characters that aren't a
part of a valid note from being entered. (I.e., anything that
isn't A-G, a number, sharp, flat, or dash). However, although the
8th and 9th octaves aren't currently supported as valid note
octaves, the characters "8" and "9" _were_ allowed to be entered
(which would cause those notes to be displayed with an error
style). This commit prevents those two characters from being
entered, for consistency with other invalid characters.